% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Modelcharts.R
\name{GAIN_CHART}
\alias{GAIN_CHART}
\title{Functions
Gain Chart}
\usage{
GAIN_CHART(INPUT, Probability, cutoffs, Outcome, Event)
}
\arguments{
\item{INPUT}{Input data}

\item{Probability}{Probability values between zero and one}

\item{cutoffs}{probability cutoffs(c(0.80,0.60,0.40,0.20,0)/c(0.5,0))}

\item{Outcome}{outcome variable(target variable)}

\item{Event}{outcome representation ("YES"/"Y"/"1")}
}
\value{
A gain chart
}
\description{
Creates a Gain chart.
}
\examples{
\dontrun{
# Run it and see for yourself
}
data.tmp<-read.csv(system.file("ext", "testdata.csv", package="Modelcharts"))
GAIN_CHART(data.tmp,data.tmp$Probability,seq(0.95,0,-0.05),data.tmp$Outcome,"Y")
}
\seealso{
\code{\link{Modelcharts}}
}
