% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Modelcharts.R
\docType{package}
\name{Modelcharts}
\alias{Modelcharts}
\alias{Modelcharts-package}
\title{Gain Chart and Lift Chart     \cr}
\description{
This Package provides two important functions for producing Gain chart and Lift chart for any classification model.
}
\section{GAIN_CHART()}{

Creates a gain chart based on calculated probability values and actual outcome.
}

\section{LIFT_CHART()}{

creates a lift chart based on calculated probability values and actual outcome.
}

\seealso{
\code{\link{GAIN_CHART}}, \code{\link{LIFT_CHART}}
}
